requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
certifi==2022.12.7
    # via pyproj
diskcache==5.4.0
    # via -r requirements.in
jinja2==3.1.2
    # via -r requirements.in
markupsafe==2.1.2
    # via jinja2
overpy==0.6
    # via -r requirements.in
pyproj==3.4.1
    # via -r requirements.in
python-slugify==8.0.1
    # via -r requirements.in
text-unidecode==1.3
    # via python-slugify
